requirements.txt
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
docutils==0.22
fonttools==4.59.1
grpcio==1.75.1
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
juliandate==1.0.5
keyring==25.6.0
kiwisolver==1.4.9
markdown-it-py==3.0.0
matplotlib==3.10.5
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.3.0
numpy==2.3.2
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
readme_renderer==44.0
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
six==1.17.0
twine==6.1.0
typing_extensions==4.15.0
urllib3==2.5.0
